chore: remove obsolete rules_schema module
